Extract Room, User to separate files from client
